Deduplicate test setup and tune Sonar exclusions
Extract registered_issuer() and nonzero_nullifier() helpers to
eliminate repeated boot+register blocks across transfer-hook smoke
tests. Exclude scripts, circuits, docs, migrations, and build
artifacts from SonarCloud analysis.
