+++ backend/programs/zksettle/tests/helpers/mod.rs
@@ -1,7 +1,7 @@
 pub mod harness;
 pub mod instructions;
 
-pub use harness::{boot_harness, funded_authority};
+pub use harness::{boot_harness, funded_authority, nonzero_nullifier, registered_issuer};
 pub use instructions::{
     create_token2022_mint_with_hook_ixs, default_light_args, execute_hook_ix, extra_meta_pda,
     hook_payload_pda, init_extra_meta_ix, issuer_pda, register_ix, set_hook_payload_ix, update_ix,
