+++ sonar-project.properties
@@ -12,3 +12,15 @@ sonar.organization=yuribodo
 
 # Encoding of the source code. Default is default system encoding
 #sonar.sourceEncoding=UTF-8
+
+# Exclude non-production code from analysis
+sonar.exclusions=\
+  scripts/**/*,\
+  circuits/**/*,\
+  docs/**/*,\
+  backend/migrations/**/*,\
+  **/target/**/*,\
+  **/*.lock
+
+# Exclude CLI scripts from coverage analysis
+sonar.coverage.exclusions=scripts/**/*
